@@ -8,24 +8,24 @@ homepage: https://github.com/flutter/cocoon
 publish_to: none
 
 environment:
-  sdk: '>=2.19.0-0 <4.0.0'
+  sdk: '>=3.0.0-0 <4.0.0'
 
 dependencies:
-  appengine: 0.13.5
+  appengine: 0.13.6
   args: 2.4.2
-  collection: 1.17.2
+  collection: 1.18.0
   corsac_jwt: 1.0.0-nullsafety.1
   crypto: 3.0.3
   dbcrypt: 2.0.0
   file: 6.1.4
   fixnum: 1.1.0
-  gcloud: 0.8.9
-  github: 9.15.1
-  googleapis: 11.2.0
+  gcloud: 0.8.10
+  github: 9.17.0
+  googleapis: 11.3.0
   googleapis_auth: 1.4.1
   gql: 0.14.0
   graphql: 5.1.3
-  grpc: 3.2.2
+  grpc: 3.2.3
   http: 0.13.6
   json_annotation: 4.8.1
   logging: 1.2.0
