@@ -6,13 +6,11 @@ import 'dart:convert';
 
 import 'package:cocoon_service/src/model/luci/buildbucket.dart';
 import 'package:meta/meta.dart';
-import 'package:retry/retry.dart';
 
 import '../../cocoon_service.dart';
 import '../model/appengine/task.dart';
 import '../request_handling/subscription_handler.dart';
 import '../service/datastore.dart';
-import '../service/exceptions.dart';
 import '../service/logging.dart';
 
 /// TODO(drewroengoogle): Make this subscription generic so we can accept more
@@ -33,49 +31,60 @@ class DartInternalSubscription extends SubscriptionHandler {
     super.authProvider,
     required this.buildBucketClient,
     @visibleForTesting this.datastoreProvider = DatastoreService.defaultProvider,
-    this.retryOptions = Config.buildbucketRetry,
   }) : super(subscriptionName: 'dart-internal-build-results-sub');
 
   final BuildBucketClient buildBucketClient;
   final DatastoreServiceProvider datastoreProvider;
-  final RetryOptions retryOptions;
 
   @override
   Future<Body> post() async {
     final DatastoreService datastore = datastoreProvider(config.db);
 
-    // This message comes from the engine_v2 recipes once a build run on
-    // dart-internal has completed.
-    //
-    // Example: https://flutter.googlesource.com/recipes/+/c6af020f0f22e392e30b769a5ed97fadace308fa/recipes/engine_v2/engine_v2.py#185
-    log.info("Getting buildbucket id from pubsub message");
-    final dynamic messageJson = json.decode(message.data.toString());
+    if (message.data == null) {
+      log.info('no data in message');
+      return Body.empty;
+    }
 
-    final int buildbucketId = messageJson['buildbucket_id'];
-    log.info("Buildbucket id: $buildbucketId");
+    final dynamic buildData = json.decode(message.data!);
+    log.info("Build data json: $buildData");
 
-    log.info("Creating build request object");
+    if (buildData["build"] == null) {
+      log.info('no build information in message');
+      return Body.empty;
+    }
+
+    final String project = buildData['build']['builder']['project'];
+    final String bucket = buildData['build']['builder']['bucket'];
+    final String builder = buildData['build']['builder']['builder'];
+
+    // This should already be covered by the pubsub filter, but adding an additional check
+    // to ensure we don't process builds that aren't from dart-internal/flutter.
+    if (project != 'dart-internal' || bucket != 'flutter') {
+      log.info("Ignoring build not from dart-internal/flutter bucket");
+      return Body.empty;
+    }
+
+    // Only publish the parent release_builder builds to the datastore.
+    // TODO(drewroengoogle): Determine which builds we want to save to the datastore and check for them.
+    final regex = RegExp(r'(Linux|Mac|Windows)\s+(engine_release_builder|packaging_release_builder)');
+    if (!regex.hasMatch(builder)) {
+      log.info("Ignoring builder that is not a release builder");
+      return Body.empty;
+    }
+
+    final String buildbucketId = buildData['build']['id'];
+    log.info("Creating build request object with build id $buildbucketId");
     final GetBuildRequest request = GetBuildRequest(
-      id: buildbucketId.toString(),
-      fields:
-          "id,builder,number,createdBy,createTime,startTime,endTime,updateTime,status,input.properties,input.gitilesCommit",
+      id: buildbucketId,
     );
 
     log.info(
       "Calling buildbucket api to get build data for build $buildbucketId",
     );
-    final Build build = await _getBuildFromBuildbucket(request);
-
-    // This is for handling subbuilds, based on the engine_v2 strategy of running
-    // builds under the same builder ({Platform} Engine Drone).
-    String? name;
-    if (build.input?.properties != null && build.input?.properties?["build"] != null) {
-      final Map<String, Object> buildProperties = build.input?.properties?["build"] as Map<String, Object>;
-      name = buildProperties["name"] as String;
-    }
+    final Build build = await buildBucketClient.getBuild(request);
 
     log.info("Checking for existing task in datastore");
-    final Task? existingTask = await datastore.getTaskFromBuildbucketBuild(build, customName: name);
+    final Task? existingTask = await datastore.getTaskFromBuildbucketBuild(build);
 
     late Task taskToInsert;
     if (existingTask != null) {
@@ -84,36 +93,12 @@ class DartInternalSubscription extends SubscriptionHandler {
       taskToInsert = existingTask;
     } else {
       log.info("Creating Task from Buildbucket result");
-      taskToInsert = await Task.fromBuildbucketBuild(build, datastore, customName: name);
+      taskToInsert = await Task.fromBuildbucketBuild(build, datastore);
     }
 
     log.info("Inserting Task into the datastore: ${taskToInsert.toString()}");
     await datastore.insert(<Task>[taskToInsert]);
 
     return Body.forJson(taskToInsert.toString());
-  }
-
-  Future<Build> _getBuildFromBuildbucket(
-    GetBuildRequest request,
-  ) async {
-    return retryOptions.retry(
-      () async {
-        final Build build = await buildBucketClient.getBuild(request);
-
-        if (build.status != Status.success &&
-            build.status != Status.failure &&
-            build.status != Status.infraFailure &&
-            build.status != Status.canceled) {
-          log.info(
-            "Build is not finished",
-          );
-          throw UnfinishedBuildException(
-            "Build is not finished",
-          );
-        }
-        return build;
-      },
-      retryIf: (Exception e) => e is UnfinishedBuildException,
-    );
   }
 }
