@@ -10,7 +10,6 @@ import 'package:cocoon_service/src/model/luci/push_message.dart' as push;
 import 'package:cocoon_service/src/service/datastore.dart';
 import 'package:gcloud/db.dart';
 import 'package:mockito/mockito.dart';
-import 'package:retry/retry.dart';
 import 'package:test/test.dart';
 
 import '../src/datastore/fake_config.dart';
@@ -31,12 +30,22 @@ void main() {
   final DateTime endTime = DateTime(2023, 1, 1, 0, 14, 23);
   const String project = "dart-internal";
   const String bucket = "flutter";
-  const String builder = "Mac amazing_builder_tests";
-  const String propertiesBuildName = "Mac Unit_Test";
+  const String builder = "Linux packaging_release_builder";
   const int buildId = 123456;
   const String fakeHash = "HASH12345";
   const String fakeBranch = "test-branch";
-  final String fakePubsubMessage = "{ \"buildbucket_id\": ${buildId.toString()} }";
+  const String fakePubsubMessage = """
+    {
+      "build": {
+        "id": "$buildId",
+        "builder": {
+          "project": "$project",
+          "bucket": "$bucket",
+          "builder": "$builder"
+        }
+      }
+    }
+  """;
 
   setUp(() async {
     config = FakeConfig();
@@ -47,7 +56,6 @@ void main() {
       authProvider: FakeAuthenticationProvider(),
       buildBucketClient: buildBucketClient,
       datastoreProvider: (DatastoreDB db) => DatastoreService(config.db, 5),
-      retryOptions: const RetryOptions(maxAttempts: 3, delayFactor: Duration.zero),
     );
     request = FakeHttpRequest();
 
@@ -77,9 +85,6 @@ void main() {
           hash: fakeHash,
           ref: "refs/heads/$fakeBranch",
         ),
-        properties: <String, Object>{
-          'build': <String, String>{'name': propertiesBuildName},
-        },
       ),
     );
     when(
@@ -94,201 +99,13 @@ void main() {
   });
 
   test('creates a new task successfully', () async {
-    tester.message = push.PushMessage(data: fakePubsubMessage);
+    tester.message = const push.PushMessage(data: fakePubsubMessage);
 
     await tester.post(handler);
 
     verify(
       buildBucketClient.getBuild(any),
     ).called(1);
-
-    // This is used for testing to pull the data out of the "datastore" so that
-    // we can verify what was saved.
-    late Task taskInDb;
-    late Commit commitInDb;
-    config.db.values.forEach((k, v) {
-      if (v is Task && v.buildNumberList == buildId.toString()) {
-        taskInDb = v;
-      }
-      if (v is Commit) {
-        commitInDb = v;
-      }
-    });
-
-    // Ensure the task has the correct parent and commit key
-    expect(
-      commitInDb.id,
-      equals(taskInDb.commitKey?.id),
-    );
-
-    expect(
-      commitInDb.id,
-      equals(taskInDb.parentKey?.id),
-    );
-
-    // Ensure the task in the db is exactly what we expect
-    final Task expectedTask = Task(
-      attempts: 1,
-      buildNumber: buildId,
-      buildNumberList: buildId.toString(),
-      builderName: builder,
-      commitKey: commitInDb.key,
-      createTimestamp: startTime.millisecondsSinceEpoch,
-      endTimestamp: endTime.millisecondsSinceEpoch,
-      luciBucket: bucket,
-      name: propertiesBuildName,
-      stageName: "dart-internal",
-      startTimestamp: startTime.millisecondsSinceEpoch,
-      status: "Succeeded",
-      key: commit.key.append(Task),
-      timeoutInMinutes: 0,
-      reason: '',
-      requiredCapabilities: [],
-      reservedForAgentId: '',
-    );
-
-    expect(
-      taskInDb.toString(),
-      equals(expectedTask.toString()),
-    );
-  });
-
-  test('updates an existing task successfully', () async {
-    const int existingTaskId = 123;
-    final Task fakeTask = Task(
-      attempts: 1,
-      buildNumber: existingTaskId,
-      buildNumberList: existingTaskId.toString(),
-      builderName: builder,
-      commitKey: commit.key,
-      createTimestamp: startTime.millisecondsSinceEpoch,
-      endTimestamp: endTime.millisecondsSinceEpoch,
-      luciBucket: bucket,
-      name: propertiesBuildName,
-      stageName: "dart-internal",
-      startTimestamp: startTime.millisecondsSinceEpoch,
-      status: "Succeeded",
-      key: commit.key.append(Task),
-      timeoutInMinutes: 0,
-      reason: '',
-      requiredCapabilities: [],
-      reservedForAgentId: '',
-    );
-    final List<Task> datastoreCommit = <Task>[fakeTask];
-    await config.db.commit(inserts: datastoreCommit);
-
-    tester.message = push.PushMessage(data: fakePubsubMessage);
-
-    await tester.post(handler);
-
-    verify(
-      buildBucketClient.getBuild(any),
-    ).called(1);
-
-    // This is used for testing to pull the data out of the "datastore" so that
-    // we can verify what was saved.
-    final String expectedBuilderList = "${existingTaskId.toString()},${buildId.toString()}";
-    late Task taskInDb;
-    late Commit commitInDb;
-    config.db.values.forEach((k, v) {
-      if (v is Task && v.buildNumberList == expectedBuilderList) {
-        taskInDb = v;
-      }
-      if (v is Commit) {
-        commitInDb = v;
-      }
-    });
-
-    // Ensure the task has the correct parent and commit key
-    expect(
-      commitInDb.id,
-      equals(taskInDb.commitKey?.id),
-    );
-
-    expect(
-      commitInDb.id,
-      equals(taskInDb.parentKey?.id),
-    );
-
-    // Ensure the task in the db is exactly what we expect
-    final Task expectedTask = Task(
-      attempts: 2,
-      buildNumber: buildId,
-      buildNumberList: expectedBuilderList,
-      builderName: builder,
-      commitKey: commitInDb.key,
-      createTimestamp: startTime.millisecondsSinceEpoch,
-      endTimestamp: endTime.millisecondsSinceEpoch,
-      luciBucket: bucket,
-      name: propertiesBuildName,
-      stageName: "dart-internal",
-      startTimestamp: startTime.millisecondsSinceEpoch,
-      status: "Succeeded",
-      key: commit.key.append(Task),
-      timeoutInMinutes: 0,
-      reason: '',
-      requiredCapabilities: [],
-      reservedForAgentId: '',
-    );
-
-    expect(
-      taskInDb.toString(),
-      equals(expectedTask.toString()),
-    );
-  });
-
-  test('retries buildbucket when build is still in progress', () async {
-    final Build fakeCompletedBuild = Build(
-      builderId: const BuilderId(project: project, bucket: bucket, builder: builder),
-      number: buildId,
-      id: 'fake-build-id',
-      status: Status.success,
-      startTime: startTime,
-      endTime: endTime,
-      input: const Input(
-        gitilesCommit: GitilesCommit(
-          project: "flutter/flutter",
-          hash: fakeHash,
-          ref: "refs/heads/$fakeBranch",
-        ),
-      ),
-    );
-
-    final Build fakeInProgressBuild = Build(
-      builderId: const BuilderId(project: project, bucket: bucket, builder: builder),
-      number: buildId,
-      id: 'fake-build-id',
-      status: Status.started,
-      startTime: startTime,
-      endTime: null,
-      input: const Input(
-        gitilesCommit: GitilesCommit(
-          project: "flutter/flutter",
-          hash: fakeHash,
-          ref: "refs/heads/$fakeBranch",
-        ),
-      ),
-    );
-
-    // The first time the mocked getBuild is called, return fakeInProgressBuild,
-    // then return fakeCompletedBuild the second time.
-    int count = 0;
-    when(
-      buildBucketClient.getBuild(
-        any,
-        buildBucketUri: "https://cr-buildbucket.appspot.com/prpc/buildbucket.v2.Builds",
-      ),
-    ).thenAnswer(
-      (_) => [Future<Build>.value(fakeInProgressBuild), Future<Build>.value(fakeCompletedBuild)][count++],
-    );
-
-    tester.message = push.PushMessage(data: fakePubsubMessage);
-
-    await tester.post(handler);
-
-    verify(
-      buildBucketClient.getBuild(any),
-    ).called(2);
 
     // This is used for testing to pull the data out of the "datastore" so that
     // we can verify what was saved.
@@ -341,24 +158,151 @@ void main() {
     );
   });
 
-  test('does not retry buildbucket when buildbucket throws an exception', () async {
-    when(
-      await buildBucketClient.getBuild(
-        any,
-        buildBucketUri: "https://cr-buildbucket.appspot.com/prpc/buildbucket.v2.Builds",
-      ),
-    ).thenThrow(
-      Exception("Failed to get from buildbucket for some reason!"),
+  test('updates an existing task successfully', () async {
+    const int existingTaskId = 123;
+    final Task fakeTask = Task(
+      attempts: 1,
+      buildNumber: existingTaskId,
+      buildNumberList: existingTaskId.toString(),
+      builderName: builder,
+      commitKey: commit.key,
+      createTimestamp: startTime.millisecondsSinceEpoch,
+      endTimestamp: endTime.millisecondsSinceEpoch,
+      luciBucket: bucket,
+      name: builder,
+      stageName: "dart-internal",
+      startTimestamp: startTime.millisecondsSinceEpoch,
+      status: "Succeeded",
+      key: commit.key.append(Task),
+      timeoutInMinutes: 0,
+      reason: '',
+      requiredCapabilities: [],
+      reservedForAgentId: '',
     );
+    final List<Task> datastoreCommit = <Task>[fakeTask];
+    await config.db.commit(inserts: datastoreCommit);
 
-    tester.message = push.PushMessage(data: fakePubsubMessage);
-    await expectLater(
-      tester.post(handler),
-      throwsA(isA<Exception>()),
-    );
+    tester.message = const push.PushMessage(data: fakePubsubMessage);
+
+    await tester.post(handler);
 
     verify(
       buildBucketClient.getBuild(any),
     ).called(1);
+
+    // This is used for testing to pull the data out of the "datastore" so that
+    // we can verify what was saved.
+    final String expectedBuilderList = "${existingTaskId.toString()},${buildId.toString()}";
+    late Task taskInDb;
+    late Commit commitInDb;
+    config.db.values.forEach((k, v) {
+      if (v is Task && v.buildNumberList == expectedBuilderList) {
+        taskInDb = v;
+      }
+      if (v is Commit) {
+        commitInDb = v;
+      }
+    });
+
+    // Ensure the task has the correct parent and commit key
+    expect(
+      commitInDb.id,
+      equals(taskInDb.commitKey?.id),
+    );
+
+    expect(
+      commitInDb.id,
+      equals(taskInDb.parentKey?.id),
+    );
+
+    // Ensure the task in the db is exactly what we expect
+    final Task expectedTask = Task(
+      attempts: 2,
+      buildNumber: buildId,
+      buildNumberList: expectedBuilderList,
+      builderName: builder,
+      commitKey: commitInDb.key,
+      createTimestamp: startTime.millisecondsSinceEpoch,
+      endTimestamp: endTime.millisecondsSinceEpoch,
+      luciBucket: bucket,
+      name: builder,
+      stageName: "dart-internal",
+      startTimestamp: startTime.millisecondsSinceEpoch,
+      status: "Succeeded",
+      key: commit.key.append(Task),
+      timeoutInMinutes: 0,
+      reason: '',
+      requiredCapabilities: [],
+      reservedForAgentId: '',
+    );
+
+    expect(
+      taskInDb.toString(),
+      equals(expectedTask.toString()),
+    );
+  });
+
+  test('ignores null message', () async {
+    tester.message = const push.PushMessage(data: null);
+    expect(await tester.post(handler), equals(Body.empty));
+  });
+
+  test('ignores message with empty build data', () async {
+    tester.message = const push.PushMessage(data: "{}");
+    expect(await tester.post(handler), equals(Body.empty));
+  });
+
+  test('ignores message not from flutter bucket', () async {
+    tester.message = const push.PushMessage(
+      data: """
+    {
+      "build": {
+        "id": "$buildId",
+        "builder": {
+          "project": "$project",
+          "bucket": "dart",
+          "builder": "$builder"
+        }
+      }
+    }
+    """,
+    );
+    expect(await tester.post(handler), equals(Body.empty));
+  });
+
+  test('ignores message not from dart-internal project', () async {
+    tester.message = const push.PushMessage(
+      data: """
+    {
+      "build": {
+        "id": "$buildId",
+        "builder": {
+          "project": "different-project",
+          "bucket": "$bucket",
+          "builder": "$builder"
+        }
+      }
+    }
+    """,
+    );
+    expect(await tester.post(handler), equals(Body.empty));
+  });
+
+  test('ignores message not from an accepted builder', () async {
+    tester.message = const push.PushMessage(
+      data: """
+    {
+      "build": {
+        "id": "$buildId",
+        "builder": {
+          "project": "different-project",
+          "bucket": "$bucket",
+          "builder": "different-builder"
+        }
+      }
+    }
+    """,
+    );
+    expect(await tester.post(handler), equals(Body.empty));
   });
 }
